Some compilation fixes and fixed tracy on different dlls creating different ports

+++ Engine/src/render/private/render_graph/render_graph_resource_aliasing2.h
@@ -11,7 +11,7 @@ namespace Mizu
 struct AliasingResource
 {
     RenderGraphResource resource;
-    uint32_t begin, end;
+    size_t begin, end;
     uint64_t size, alignment, offset;
 };
